{"types":"C","runs":49}

+++ problem16.java
@@ -4,9 +4,9 @@ public class problem16 extends ConsoleProgram {
 	
 	public void run () {
 		
-		int a = readInt ("enter first number:");
-		int b = readInt ("enter second number:");
-		int c = readInt ("enter third number:");
+		double a = readDouble ("enter first number:");
+		double b = readDouble ("enter second number:");
+		double c = readDouble ("enter third number:");
 		double averageHarmonic = 1/(1/a + 1/b + 1/c);
 		println ("The Average Harmonic is:" + averageHarmonic);
 	}
